Now it is possible to specify decimal values on geographic coordinates (this is obviously true only in the "seconds" coordinates) (Issue-739).
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@3230 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/bia/command/geobase/EditGeographicCoordinatesPlaceCommand.java
@@ -40,11 +40,11 @@ public class EditGeographicCoordinatesPlaceCommand {
 	private Integer placeGeographicCoordinatesId;
 	private Integer degreeLatitude;
 	private Integer minuteLatitude;
-	private Integer secondLatitude;
+	private Float secondLatitude;
 	private String directionLatitude;
 	private Integer degreeLongitude;
 	private Integer minuteLongitude;
-	private Integer secondLongitude;
+	private Float secondLongitude;
 	private String directionLongitude;
 
 	/**
@@ -97,14 +97,14 @@ public class EditGeographicCoordinatesPlaceCommand {
 	/**
 	 * @param secondLatitude the secondLatitude to set
 	 */
-	public void setSecondLatitude(Integer secondLatitude) {
+	public void setSecondLatitude(Float secondLatitude) {
 		this.secondLatitude = secondLatitude;
 	}
 
 	/**
 	 * @return the secondLatitude
 	 */
-	public Integer getSecondLatitude() {
+	public Float getSecondLatitude() {
 		return secondLatitude;
 	}
 
@@ -153,14 +153,14 @@ public class EditGeographicCoordinatesPlaceCommand {
 	/**
 	 * @param secondLongitude the secondLongitude to set
 	 */
-	public void setSecondLongitude(Integer secondLongitude) {
+	public void setSecondLongitude(Float secondLongitude) {
 		this.secondLongitude = secondLongitude;
 	}
 
 	/**
 	 * @return the secondLongitude
 	 */
-	public Integer getSecondLongitude() {
+	public Float getSecondLongitude() {
 		return secondLongitude;
 	}
